requirements.txt
psycopg[binary]>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0
python-dotenv>=1.0.0
